requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.10.11
aiosignal==1.3.2
altair==5.5.0
amqp==5.3.1
annotated-types==0.7.0
anyio==4.9.0
asgiref==3.8.1
assemblyai==0.37.0
attrs==25.3.0
av==14.3.0
awscli==1.38.37
beautifulsoup4==4.13.3
billiard==4.2.1
blinker==1.9.0
boto3==1.37.37
botocore==1.37.37
Brotli==1.0.9
cachetools==5.5.2
celery==5.5.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
coloredlogs==15.0.1
cron-descriptor==1.4.5
cryptography==44.0.2
ctranslate2==4.6.0
dataclasses-json==0.6.7
# decorator==5.1.1
Deprecated==1.2.18
distro==1.9.0
Django==5.1.7
django-celery-beat==2.8.0
django-timezone-field==7.1
djangorestframework==3.15.2
docutils==0.19
drf-spectacular==0.28.0
faster-whisper==1.1.1
filelock==3.18.0
filetype==1.2.0
flatbuffers==25.2.10
frozenlist==1.5.0
fsspec==2025.3.2
future==1.0.0
gevent==25.4.1
gitdb==4.0.12
GitPython==3.1.44
google-ai-generativelanguage==0.6.17
google-api-core==2.24.2
google-api-python-client==2.167.0
google-auth==2.39.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.1
googleapis-common-protos==1.70.0
greenlet==3.2.0
groq==0.20.0
grpcio==1.72.0rc1
grpcio-status==1.71.0
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httplib2==0.22.0
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
inflection==0.5.1
iniconfig==2.1.0
Jinja2==3.1.6
jmespath==1.0.1
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kombu==5.5.2
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-google-genai==2.1.2
langchain-pinecone==0.2.5
langchain-tests==0.3.17
langchain-text-splitters==0.3.8
langsmith==0.3.31
MarkupSafe==3.0.2
marshmallow==3.26.1
mkl_fft==1.3.11
mkl_random==1.2.8
mkl-service==2.4.2
moviepy==1.0.3
mpmath==1.3.0
multidict==6.4.3
mypy-extensions==1.0.0
narwhals==1.35.0
networkx==3.4.2
# numpy==2.0.1
oauthlib==3.2.2
onnxruntime==1.21.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pendulum==3.0.0
pillow==10.4.0
pinecone==6.0.2
pinecone-client==6.0.0
pinecone-plugin-interface==0.0.7
pip==25.0
pluggy==1.5.0
proglog==0.1.10
prompt_toolkit==3.0.50
propcache==0.3.1
proto-plus==1.26.1
protobuf==5.29.4
psycopg2-binary==2.9.10
pyarrow==19.0.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
pydantic==2.11.1
pydantic_core==2.33.0
pydantic-settings==2.8.1
pydeck==0.9.1
PyJWT==2.9.0
pyparsing==3.2.3
pyreadline3==3.5.4
PySocks==1.7.1
pytest==8.3.5
pytest-cov==5.0.0
pytest-asyncio==0.26.0
pytest-socket==0.7.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
PyYAML==6.0.2
redis==5.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
requests-oauthlib==2.0.0
requests-toolbelt==1.0.0
rpds-py==0.24.0
rsa==4.7.2
s3transfer==0.11.5
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.1.0
setuptools==75.8.0
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.40
sqlparse==0.5.3
streamlit==1.44.1
svix==1.64.1
sympy==1.13.1
syrupy==4.9.1
tenacity==9.1.2
threadpoolctl==3.6.0
time-machine==2.16.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.51.3
types-Deprecated==1.2.15.20250304
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.0
typing-inspect==0.9.0
typing-inspection==0.4.0
tzdata==2025.2
uritemplate==4.1.1
urllib3==2.3.0
vine==5.1.0
watchdog==6.0.0
wcwidth==0.2.13
websockets==15.0.1
wheel==0.45.1
whitenoise==6.9.0
win-inet-pton==1.1.0
wrapt==1.17.2
yarl==1.19.0
yt-dlp==2025.3.31
zope.event==5.0
zope.interface==7.2
zstandard==0.23.0
